# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation API for Hussam & Yara
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RsvpInput {
  /** @minLength 1 */
  guestName: string;
  attending: boolean;
  hasPlusOne?: boolean;
  /** @nullable */
  plusOneName?: string | null;
}

export interface Rsvp {
  id: number;
  guestName: string;
  attending: boolean;
  hasPlusOne: boolean;
  /** @nullable */
  plusOneName?: string | null;
  submittedAt: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
  /** @nullable */
  message?: string | null;
}

export interface ErrorResponse {
  error: string;
}