# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wedding invitation API for Hussam & Yara
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: rsvp
    description: RSVP operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /rsvp:
    post:
      operationId: submitRsvp
      tags: [rsvp]
      summary: Submit RSVP response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpInput"
      responses:
        "201":
          description: RSVP submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rsvp"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rsvp/all:
    get:
      operationId: getAllRsvps
      tags: [admin]
      summary: Get all RSVP submissions (admin only)
      responses:
        "200":
          description: List of all RSVP submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rsvp"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Check admin session status
      responses:
        "200":
          description: Session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RsvpInput:
      type: object
      required:
        - guestName
        - attending
      properties:
        guestName:
          type: string
          minLength: 1
        attending:
          type: boolean
        hasPlusOne:
          type: boolean
        plusOneName:
          type: ["string", "null"]

    Rsvp:
      type: object
      required:
        - id
        - guestName
        - attending
        - hasPlusOne
        - submittedAt
      properties:
        id:
          type: integer
        guestName:
          type: string
        attending:
          type: boolean
        hasPlusOne:
          type: boolean
        plusOneName:
          type: ["string", "null"]
        submittedAt:
          type: string

    AdminCredentials:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    AdminSession:
      type: object
      required:
        - authenticated
      properties:
        authenticated:
          type: boolean
        message:
          type: ["string", "null"]

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string