# lib/api-zod/src/generated/types/rsvpInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation API for Hussam & Yara
 * OpenAPI spec version: 0.1.0
 */

export interface RsvpInput {
  /** @minLength 1 */
  guestName: string;
  attending: boolean;
  hasPlusOne?: boolean;
  /** @nullable */
  plusOneName?: string | null;
}